revising plotting of leveling/insar

+++ Geodesy_Modeling/UAVSAR/visualize.py
@@ -14,8 +14,8 @@ def visualize_timedep_timeseries(config_filename, gps_filename, outdir):
     file_dict = multiSAR_utilities.get_file_dictionary(config_filename);
     uav_los, uav_lon, uav_lat = file_dict["uavsar_file"], file_dict["uavsar_lon"], file_dict["uavsar_lat"];
     myUAVSAR_TS = UAVSAR.uavsar_readwrite.inputs_TS_grd(uav_los, uav_lon, uav_lat);
-    [gps_lons, gps_lats, gps_names] = np.loadtxt(gps_filename, unpack=True, dtype={"formats":(np.float, np.float, 'U4'),
-                                                                                   "names":("lons", "lats", "name")});
+    [gps_lons, gps_lats, gps_names] = np.loadtxt(gps_filename, unpack=True, dtype={"formats": (float, float, 'U4'),
+                                                                                   "names": ("lons", "lats", "name")});
     selected_epochs = np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);  # allows you to combine intervals if necessary
     # selected currently breaks because it's a list not slices. But this is not a big deal.
     UAVSAR.uavsar_readwrite.total_ts_visualizing(myUAVSAR_TS, gps_lons, gps_lats, gps_names, selected_epochs, outdir);
